pyproject.toml
[project]
name = "high-quality-translation"
version = "0.0.0"
description = "Validates that the translation is of high quality."
authors = [
    {name = "Guardrails AI", email = "contact@guardrailsai.com"}
]
license = {file = "LICENSE"}
readme = "README.md"
requires-python = ">= 3.8"
dependencies = [
    "guardrails-ai>=0.3.2",
    "guardrails-ai-unbabel-comet"
]

[project.optional-dependencies]
dev = [
    "pyright",
    "pytest",
    "ruff"
]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-rP"
testpaths = [
    "test"
]

[tool.pyright]
include = ["validator"]